Epic E: Literal Parameterization Rewrite
